+++ services/Job1111/src/com/psabackoffice/job1111/service/Job1111ProcedureExecutorService_V1.java
@@ -34,7 +34,7 @@ public interface Job1111ProcedureExecutorService_V1 {
     List<Object> executeGetJobPCOT(Integer runMode);
 
     @Deprecated
-    List<Object> executeCreateManualActivityPipe(Integer varBidId, Integer varQuantity, String varNotes, Integer varUserId);
+    List<Object> executeCreateManualActivityPipe(Integer varBidId, BigDecimal varQuantity, String varNotes, Integer varUserId);
 
     @Deprecated
     List<Object> executeGetCivilPCOT(Integer runMode);
